config_parser-20230220T090119Z-001/config_parser/requirements.txt
PyYAML == 5.4.1
configparser == 5.0.2
pytest == 6.2.4
